+++ src/ee/ee.hpp
@@ -62,7 +62,7 @@ private:
 extern bool in_branch_delay_slot;
 extern u32 jump_addr, pc;
 extern Reg128 lo, lo1, hi, hi1;
-extern u64 sa;
+extern u32 sa;
 
 void cancel_jump();
 bool init();
